Split error module into submodules
