+++ exonum/src/runtime/error/error_match.rs
@@ -1,0 +1,158 @@
+// Copyright 2020 The Exonum Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::fmt;
+
+use super::{CallType, ErrorKind, ExecutionFail};
+
+use crate::runtime::InstanceId;
+
+/// Matcher for `ExecutionError`s that can have some fields unspecified. Can be compared to
+/// an `ExceptionError`, e.g., in tests. The unspecified fields will match any value in the error.
+///
+/// # Examples
+///
+/// ```
+/// use exonum::runtime::{ExecutionError, InstanceId, ErrorMatch};
+/// use exonum_derive::ExecutionFail;
+///
+/// #[derive(Debug, ExecutionFail)]
+/// pub enum Error {
+///     /// Content hash already exists.
+///     HashAlreadyExists = 0,
+///     // other variants...
+/// }
+///
+/// // Identifier of the service that will cause an error.
+/// const SERVICE_ID: InstanceId = 100;
+///
+/// # fn not_run(error: ExecutionError) {
+/// let err: &ExecutionError = // ...
+/// #    &error;
+/// let matcher = ErrorMatch::from_fail(&Error::HashAlreadyExists)
+///     .for_service(SERVICE_ID);
+/// assert_eq!(*err, matcher);
+/// # }
+/// ```
+#[derive(Debug)]
+pub struct ErrorMatch {
+    pub(super) kind: ErrorKind,
+    pub(super) description: StringMatch,
+    pub(super) runtime_id: Option<u32>,
+    pub(super) instance_id: Option<InstanceId>,
+    pub(super) call_type: Option<CallType>,
+}
+
+impl ErrorMatch {
+    /// Creates a matcher from the provided error.
+    ///
+    /// The converted error has a kind and description set to the values returned
+    /// by the corresponding methods of the [`ExecutionFail`] trait. The call site information
+    /// (e.g., the instance ID) is not set.
+    ///
+    /// [`ExecutionFail`]: trait.ExecutionFail.html
+    pub fn from_fail<F: ExecutionFail + ?Sized>(fail: &F) -> Self {
+        Self::new(fail.kind(), fail.description())
+    }
+
+    /// Creates a matcher for `Unexpected` kind of errors.
+    /// By default it will match any description.
+    pub fn any_unexpected() -> Self {
+        Self {
+            kind: ErrorKind::Unexpected,
+            description: StringMatch::Any,
+            runtime_id: None,
+            instance_id: None,
+            call_type: None,
+        }
+    }
+
+    pub(super) fn new(kind: ErrorKind, description: String) -> Self {
+        Self {
+            kind,
+            description: StringMatch::Exact(description),
+            runtime_id: None,
+            instance_id: None,
+            call_type: None,
+        }
+    }
+
+    /// Accepts an error with any description.
+    pub fn with_any_description(mut self) -> Self {
+        self.description = StringMatch::Any;
+        self
+    }
+
+    /// Accepts an error with any description containing the specified string.
+    pub fn with_description_containing(mut self, pat: impl Into<String>) -> Self {
+        self.description = StringMatch::Contains(pat.into());
+        self
+    }
+
+    /// Accepts an error with any description matching the specified closure.
+    pub fn with_description_matching<P>(mut self, pat: P) -> Self
+    where
+        P: Fn(&str) -> bool + 'static,
+    {
+        self.description = StringMatch::Generic(Box::new(pat));
+        self
+    }
+
+    /// Accepts an error that has occurred in a runtime with the specified ID.
+    pub fn in_runtime(mut self, runtime_id: u32) -> Self {
+        self.runtime_id = Some(runtime_id);
+        self
+    }
+
+    /// Accepts an error that has occurred in a service with the specified ID.
+    pub fn for_service(mut self, instance_id: InstanceId) -> Self {
+        self.instance_id = Some(instance_id);
+        self
+    }
+
+    /// Accepts an error that has occurred in a call of the specified format.
+    pub fn in_call(mut self, call_type: CallType) -> Self {
+        self.call_type = Some(call_type);
+        self
+    }
+}
+
+pub(super) enum StringMatch {
+    Any,
+    Exact(String),
+    Contains(String),
+    Generic(Box<dyn Fn(&str) -> bool>),
+}
+
+impl fmt::Debug for StringMatch {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            StringMatch::Any => formatter.write_str("Any"),
+            StringMatch::Exact(s) => formatter.debug_tuple("Exact").field(s).finish(),
+            StringMatch::Contains(s) => formatter.debug_tuple("Contains").field(s).finish(),
+            StringMatch::Generic(_) => formatter.debug_tuple("Generic").field(&"_").finish(),
+        }
+    }
+}
+
+impl StringMatch {
+    pub(super) fn matches(&self, s: &str) -> bool {
+        match self {
+            StringMatch::Any => true,
+            StringMatch::Exact(ref_str) => ref_str == s,
+            StringMatch::Contains(needle) => s.contains(needle),
+            StringMatch::Generic(match_fn) => match_fn(s),
+        }
+    }
+}
