+++ exonum/src/runtime/mod.rs
@@ -145,7 +145,7 @@ pub use self::{
     dispatcher::{Dispatcher, Error as DispatcherError, Mailbox, Schema as DispatcherSchema},
     error::{
         catch_panic, CallSite, CallType, ErrorCode, ErrorKind, ErrorMatch, ExecutionError,
-        ExecutionFail, ExecutionStatus, SerdeExecutionStatus,
+        ExecutionFail, ExecutionStatus,
     },
     types::{
         AnyTx, ArtifactId, ArtifactSpec, ArtifactState, ArtifactStatus, CallInfo, InstanceId,
@@ -155,7 +155,8 @@ pub use self::{
 
 // Re-export for serializing `ExecutionError` via `serde`.
 #[doc(hidden)]
-pub use error::execution_error as execution_error_serde;
+pub use error::execution_error::ExecutionErrorSerde;
+
 #[macro_use]
 pub mod rust;
 pub mod migrations;
